+++ toolkit/components/filewatcher/NativeFileWatcherCommons.h
@@ -17,33 +17,10 @@
 #include <queue>
 #include <vector>
 
-#include <CoreServices/CoreServices.h>
-#include <CoreServices/Components.k.h>
 
 namespace mozilla {
 
 namespace moz_filewatcher {
-
-struct CallBackAction{
-    CallBackAction(char*eventPath, const FSEventStreamEventFlags eventFlag, const FSEventStreamEventId eventIds)
-        : mEventFlags(eventFlag)
-        , mEventIds(eventIds)
-    {
-        snprintf(mEventPath, 2048, "%s", eventPath);
-    }
-
-    char mEventPath[2048]; // FIXME: find max path length
-    const FSEventStreamEventFlags mEventFlags;
-    const FSEventStreamEventId mEventIds;
-};
-
-struct CallBackEvents{
-    CallBackEvents()
-        : callBackLock("filewatcher::eventLock")
-    {}
-    std::queue<CallBackAction> mSavedEvents;
-    Mutex callBackLock;
-};
 
 /**
  * An event used to notify the main thread when an error happens.
